+++ Flow/src/utopia/flow/collection/immutable/caching/cache/Cache.scala
@@ -4,10 +4,32 @@ import utopia.flow.collection.template.MapAccess
 
 import scala.collection.mutable
 import scala.concurrent.ExecutionContext
-import scala.concurrent.duration.{Duration, FiniteDuration}
+import scala.concurrent.duration.FiniteDuration
 
 object Cache
 {
+	// COMPUTED -------------------------
+	
+	/**
+	  * Access to weak cache constructors
+	  */
+	def weak = WeakCache
+	/**
+	  * @return Access to expiring cache constructors
+	  */
+	def expiring = ExpiringCache
+	/**
+	  * @return Access to releasing cache constructors
+	  */
+	def releasing = ReleasingCache
+	/**
+	  * @return Access to cache latest -constructors
+	  */
+	def onlyLatest = CacheLatest
+	
+	
+	// OTHER    -------------------------
+	
 	/**
 	  * Creates a new cache
 	  * @param request A function for retrieving the cached value
@@ -18,27 +40,6 @@ object Cache
 	def apply[Key, Value](request: Key => Value): Cache[Key, Value] = new _Cache[Key, Value](request)
 	
 	/**
-	  * Creates a weakly referencing cache
-	  * @param request Function for requesting new values
-	  * @tparam K Type of keys used
-	  * @tparam V Type of values used
-	  * @return A new cache
-	  */
-	def weak[K, V <: AnyRef](request: K => V) = WeakCache(request)
-	
-	/**
-	  * Creates a cache that removes cached items after a while
-	  * @param request Function for requesting new values
-	  * @param calculateExpiration A function for calculating item expiration time
-	  * @tparam K Type of keys used
-	  * @tparam V Type of values used
-	  * @return A new cache
-	  */
-	def expiring[K, V](request: K => V)(calculateExpiration: (K, V) => Duration)
-	                  (implicit exc: ExecutionContext) =
-		ExpiringCache(request)(calculateExpiration)
-	
-	/**
 	  * Creates a temporarily caching cache
 	  * @param threshold A time threshold after which items are removed
 	  * @param request Function for requesting new values
@@ -46,22 +47,10 @@ object Cache
 	  * @tparam V Type of values used
 	  * @return A new cache
 	  */
+    @deprecated("Please use .expiring.after(...) instead", "v2.5")
 	def expiringAfter[K, V](threshold: FiniteDuration)(request: K => V)
 	                       (implicit exc: ExecutionContext) =
 		ExpiringCache.after(threshold)(request)
-	
-	/**
-	  * Creates a cache that uses weak references after a while
-	  * @param request Function for requesting new values
-	  * @param calculateReferenceLength A function for calculating the length of time to use for
-	  *                                 holding a strong reference
-	  * @tparam K Type of keys used
-	  * @tparam V Type of values used
-	  * @return A new cache
-	  */
-	def releasing[K, V <: AnyRef](request: K => V)(calculateReferenceLength: (K, V) => Duration)
-	                             (implicit exc: ExecutionContext) =
-		ReleasingCache(request)(calculateReferenceLength)
 	
 	/**
 	  * Creates a cache that uses weak references after a while
@@ -71,19 +60,11 @@ object Cache
 	  * @tparam V Type of values used
 	  * @return A new cache
 	  */
+	@deprecated("Please use .releasing.after(...) instead", "v2.5")
 	def releasingAfter[K, V <: AnyRef](threshold: FiniteDuration)
 	                                  (request: K => V)
 	                                  (implicit exc: ExecutionContext) =
 		ReleasingCache.after(threshold)(request)
-	
-	/**
-	  * Creates a new cache that only caches the latest requested value
-	  * @param f A function for generating values for requested keys
-	  * @tparam K Type of keys used
-	  * @tparam V Type of values returned
-	  * @return A new cache
-	  */
-	def onlyLatest[K, V](f: K => V) = CacheLatest(f)
 		
 	
 	// NESTED   ---------------------------